added comments for all important and slight refactor/rename
