+++ make_celery.py
@@ -1,4 +1,5 @@
 from website import create_app
 
+#File used to configure celery workers
 flask_app = create_app()
 celery_app = flask_app.extensions["celery"]
