+++ website/lines/templates/arbitrage.html
@@ -10,7 +10,7 @@
 </head>
 <body>
     <br>
-    <form class="search-bar" method="get" action="{{ url_for('lines.arbitrage_opportunities') }}">
+    <form class="search-bar" method="get" action="{{ url_for('lines.search_arbitrage') }}">
         <div class="form-group">
             <label for="date">Date:</label>
             <input type="date" id="date" name="date">
@@ -52,7 +52,7 @@
                         {% for page_num in pagination.iter_pages() %}
                             {% if page_num %}
                                 {% if page_num != pagination.page %}
-                                <li><a href="{{ url_for('lines.arbitrage_opportunities', page=page_num, date=request.args.get('date'), team1=request.args.get('team1'), team2=request.args.get('team2')) }}">{{ page_num }}</a></li>
+                                <li><a href="{{ url_for('lines.search_arbitrage', page=page_num, date=request.args.get('date'), team1=request.args.get('team1'), team2=request.args.get('team2')) }}">{{ page_num }}</a></li>
                                 {% else %}
                                     <li class='active'>{{ page_num }}</li>
                                 {% endif %}
