+++ website/lines/templates/games.html
@@ -72,7 +72,7 @@
                             {% endfor %}
                         </ul>
                     {% endif %}
-                    <form method="post" action="{{ url_for('lines.call_insert_scores') }}">
+                    <form method="post" action="{{ url_for('lines.call_insert_data') }}">
                         <input type="submit" value="Fetch new data">
                     </form>
                 </div>
