+++ website/users/static/profile.js
@@ -1,3 +1,4 @@
+//Function to toggle follow/unfollow when visiting a different user's profile
 function toggleFollow(username) {
     fetch("/follow/"+username, {
     method: "POST",
